pyproject.toml
[project]
name = "dash-mcp-showcase"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dash (>=3.1.1,<4.0.0)",
    "dash-bootstrap-components (>=2.0.3,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "openai (>=1.97.0,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "google-cloud-secret-manager (>=2.20.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "dash_mcp_showcase", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
